Update: Add rauc hooks to install a bootloader update (minimize writes to boot)
This update removes the rauc-mark-good and boot retry handling from the boot
process. It's intention is to minimize write access to the boot partition of
the pi. The bootpartition is formated with a VFAT filesystem and therefor may
cause problems on a power loss, if some data was not fully written to the
emmc.

It was agreed with conti, etr ibiss and ks-system to not use the rauc
fallback handling as intended and instead only use rauc to install the
updates to the right boot slot and swith to the just installed boot slot
after successfull installation.

In the eyes of ks-system this shall help to reduce the risk of broken
sd-cards.